chore(S3FileSystemProvider): move `pathIteratorForPublisher` and related functions to `S3DirectoryStream`
